% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kms.R
\name{bw_kms}
\alias{bw_kms}
\title{Compute the bandwidth}
\usage{
bw_kms(x, type = c("hpi_linear_s1", "rot_up")[2])
}
\arguments{
\item{x.}{A matrix of size \code{c(nx, d + 1)} with the initial points.}

\item{type.}{The specific way of calculating the bandwidth: hpi linear or rot up.
Defaults to \code{"rot_up"}.}
}
\description{
Calculate the bandwidth associated for the given parameters.
}
