% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clusterFactory}
\alias{clusterFactory}
\title{Cluster factory}
\usage{
clusterFactory(num_cores, outfile = "")
}
\arguments{
\item{num_cores}{number of cores to be available in the cluster.}

\item{outfile}{text file where print the output, empty means no output.}
}
\value{
the cluster created.
}
\description{
Cluster factory for concurrent tasks.
}
\examples{
clusterFactory(2)
clusterFactory(2, "log.txt")
file.remove("log.txt")
}
