% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cos_sim_i}
\alias{cos_sim_i}
\title{Cosine similarities vector for the i-th observation}
\usage{
cos_sim_i(cos_sim_psh, i, r, n)
}
\arguments{
\item{cos_sim_psh}{cosine similarities array of dimension \code{c(n, n, r)} for the polysphere.}

\item{i}{corresponds to the \eqn{i}-th observation index.}

\item{r}{the number of spheres in the polysphere}

\item{n}{the sample size.}
}
\description{
Obtain the index of an element within a vector for the equivalent in
the upper triangular matrix without diagonal
}
\examples{
n <- 6
r <- 6
x <- sphunif::r_unif_sph(n, 3, r)
cos_sim_psh <- drop(sphunif::Psi_mat(x, scalar_prod = TRUE))
cos_sim_i(cos_sim_psh, 4, r, n)
}
