% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dimension.R
\name{cosine_polysph}
\alias{cosine_polysph}
\title{Calculates the cosine similarity for each matrix}
\usage{
cosine_polysph(x)
}
\arguments{
\item{x}{array 3-dimensional}
}
\value{
a 3-dimensional array with the cosine similarities of each matrix of \code{x}
}
\description{
Calculates the cosine similarity for each matrix
}
\examples{
x <- gen_polysphere(100, 2, 3)
cosine_polysph(x)
}
