% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_distribution.R
\name{d_i_psph_cauchy}
\alias{d_i_psph_cauchy}
\title{Marginal poly-spherical Cauchy density for the \eqn{i}-th observation}
\usage{
d_i_psph_cauchy(x, i, rho_list)
}
\arguments{
\item{x}{an array of size \code{c(n, d + 1, r)} with the poly-spherical data, where \code{n} is the number of observations, \code{d} is the dimension of each sphere, and \code{r} is the number of spheres.}

\item{i}{corresponds to the \eqn{i}-th observation index}

\item{rho_list}{rho list of size \code{n} for each \code{i}-th observation that stands for the concentration parameter.}
}
\value{
poly-spherical Cauchy density value given the parameters
}
\description{
Calculate the high-dimension poly-spherical Cauchy density for the
\eqn{i}-th observation.
}
\examples{
x <- sphunif::r_unif_sph(20, 3, 4)
d_i_psph_cauchy(x, 1, rep(0.5, 20))
d_i_psph_cauchy(x, 4, rep(0.9999, 20))
}
