% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_distribution.R
\name{d_i_sph_cauchy}
\alias{d_i_sph_cauchy}
\title{Marginal spherical Cauchy density function}
\usage{
d_i_sph_cauchy(y, i, rho)
}
\arguments{
\item{y}{matrix that stands for an sphere, \eqn{\mathcal{S}^d}.}

\item{i}{corresponds to the \eqn{i}-th observation index.}

\item{rho}{concentration parameter must be in [0, 1).}
}
\value{
Spherical marginal Cauchy density probability given the parameters.
}
\description{
Calculate the marginal low-dimension spherical
Cauchy probability.
}
