% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_distribution.R
\name{d_psph_cauchy}
\alias{d_psph_cauchy}
\title{Poly-spherical Cauchy density for the \eqn{i}-th and \eqn{j}-th observations}
\usage{
d_psph_cauchy(x, i, j, rho)
}
\arguments{
\item{x}{an array of size \code{c(n, d + 1, r)} with the poly-spherical data, where \code{n} is the number of observations, \code{d} is the dimension of each sphere, and \code{r} is the number of spheres.}

\item{i}{corresponds to the i-th observation index}

\item{j}{corresponds to the j-th observation index}

\item{rho}{param between 0 and 1 (not included)}
}
\value{
poly-spherical Cauchy density value given the parameters
}
\description{
Calculate the high-dimension poly-spherical Cauchy density for the
\eqn{i}-th and \eqn{j}-th observations.
}
