% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_distribution.R
\name{d_sph_cauchy}
\alias{d_sph_cauchy}
\title{High-dimension poly-spherical Cauchy density}
\usage{
d_sph_cauchy(x, i, j, rho, k, p)
}
\arguments{
\item{x}{an array of size \code{c(n, d + 1, r)} with the poly-spherical data, where \code{n} is the number of observations, \code{d} is the dimension of each sphere, and \code{r} is the number of spheres.}

\item{i}{corresponds to the i-th observation index.}

\item{j}{corresponds to the j-th observation index.}

\item{rho}{param between 0 and 1 (not included).}

\item{k}{corresponds to the k-th sphere index.}

\item{p}{corresponds to S^p, associated to R^(p+1).}
}
\value{
spherical Cauchy density value given the parameters.
}
\description{
Calculate the high-dimension spherical Cauchy density function.
}
