% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dimension.R
\name{diag_3d}
\alias{diag_3d}
\title{Set a specific value to the diagonal of a matrix}
\usage{
diag_3d(x, k, val)
}
\arguments{
\item{x}{array 3-dimensional}

\item{k}{the k-th matrix of the 3d array}

\item{val}{the specific value to set in the diagonal}
}
\value{
The \code{k}-th matrix of \code{x} with the diagonal set to \code{val}
}
\description{
Set a specific value to the diagonal of a matrix
}
\examples{
x <- gen_polysphere(100, 2, 3)
diag_3d(x, 1, 0)
diag_3d(x, 3, 0)
}
