% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kms.R
\name{eta_hat}
\alias{eta_hat}
\title{Estimation fo the eta parameter}
\usage{
eta_hat(x, h, data)
}
\arguments{
\item{x}{a matrix of size \code{c(nx, d + 1)} with the initial points.}

\item{h}{bandwith parameter}

\item{data}{a matrix of size \code{c(n, d + 1)} with the sample.}
}
\value{
estimation of the eta parameter.
}
\description{
Auxiliar function to run kernel mean shift clustering with scalar values.
Calculate the estimation of the parameter eta.
}
