% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kms.R
\name{euler}
\alias{euler}
\title{Compute the euler value}
\usage{
euler(x, N = 1e+06, verbose = FALSE, epsilon = 1e-05, eta_hat_efic)
}
\arguments{
\item{x}{a matrix of size \code{c(nx, d + 1)} with the initial points.}

\item{eta_hat_efic}{eta parameter value.}

\item{N.}{Defaults to \code{1e6}.}

\item{verbose.}{Defaults to \code{FALSE}.}

\item{epsilon.}{Defaults to \code{1e-5}.}
}
\value{
euler value for the given parameters.
}
\description{
Auxiliar function to run kernel mean shift clustering with scalar values.
Calculate the estimation of the euler value.
}
