% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fibonacci_lattice}
\alias{fibonacci_lattice}
\title{Fibonacci lattice algorithm to generate evenly separated points}
\usage{
fibonacci_lattice(n)
}
\arguments{
\item{n}{positive integer with the size of the grid to generate.}
}
\value{
Evenly optimal separated points onto the low-dimension sphere
\eqn{\mathcal{S}^2}.
}
\description{
Generated optimal evenly separated points onto the sphere
\eqn{\mathcal{S}^2}.
}
\examples{
fibonacci_lattice(100)
fibonacci_lattice(250)
}
\keyword{internal}
