% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_opt_sphere}
\alias{gen_opt_sphere}
\title{Generate optimum evenly separated points}
\usage{
gen_opt_sphere(n, d)
}
\arguments{
\item{n}{positive integer that defines the size of the sample to generate.}

\item{d}{size of the low-dimension which defines the sphere \eqn{\mathcal{S}^d}.}
}
\value{
evenly optimal separated points onto the low-dimension sphere \eqn{\mathcal{S}^d}.
}
\description{
Generated optimal evenly separated points onto the sphere \eqn{\mathcal{S}^d}.
}
\examples{
gen_opt_sphere(100, 1)
gen_opt_sphere(250, 2)
}
