% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_polysphere}
\alias{gen_polysphere}
\title{Poly-spherical data generation}
\usage{
gen_polysphere(n, p, r)
}
\arguments{
\item{n}{sample size}

\item{p}{number of dimension in the sphere (S^p)}

\item{r}{number of spheres (S^p)^r}
}
\value{
data onto the sphere
}
\description{
Poly-spherical data generation
}
\examples{
gen_polysphere(100, 1, 3)
gen_polysphere(250, 2, 5)
}
