% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grid_sphere}
\alias{grid_sphere}
\title{Generate optimum evenly separated points}
\usage{
grid_sphere(n, d)
}
\arguments{
\item{n}{positive integer with the size of the grid to generate.}

\item{d}{size of the low-dimension which defines the sphere
\eqn{\mathcal{S}^d}, must be 1 or 2.}
}
\value{
For \eqn{\mathcal{S}^1}, evenly optimal separated points.
For \eqn{\mathcal{S}^2}, Fibonacci lattice is applied to generate points.
\eqn{\mathcal{S}^d}.
}
\description{
Generated optimal evenly separated points on
\eqn{\mathcal{S}^1} or \eqn{\mathcal{S}^2}.
}
\examples{
grid_sphere(100, 1)
grid_sphere(250, 2)
}
