% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dimension.R
\name{high_dim_mat}
\alias{high_dim_mat}
\title{Polyspherical Cauchy conditional probability matrix (matrix version)}
\usage{
high_dim_mat(x, rho_list, cos_psh = NULL)
}
\arguments{
\item{x}{an array of size \code{c(n, d + 1, r)} with the polyspherical data, where \code{n} is the number of observations, \code{d} is the dimension of each sphere, and \code{r} is the number of spheres.}

\item{rho_list}{rho list of size \code{n} for each \code{i}-th observation that stands for the concentration parameter.}

\item{cos_psh}{cosine similarities array of dimension \code{c(n, n, r)} for the polysphere.}
}
\value{
An array of size \code{c(n, n)} with the high-dimension conditional probabilities of \code{x}.
}
\description{
Calculates the high-dimension conditional probabilities of a polyspherical Cauchy distribution. Matrix version algorithm.
}
\examples{
n <- 100
rho_list <- rep(seq(1, 10), each = n / 10)
x <- sphunif::r_unif_sph(n, 3, 3)
high_dimension_mat(x, rho_list)
high_dimension_mat(x, rho_list, cosine_polysph(x))
}
