% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dimension.R
\name{high_dimension}
\alias{high_dimension}
\title{Polyspherical Cauchy conditional probability matrix (matrix version)}
\usage{
high_dimension(x, rho_list, cos_sim_psh = NULL)
}
\arguments{
\item{x}{an array of size \code{c(n, d + 1, r)} with the polyspherical data,
where \code{n} is the number of observations, \code{d} is the dimension of
each sphere, and \code{r} is the number of spheres.}

\item{rho_list}{rho list of size \code{n} for each \code{i}-th observation
that stands for the concentration parameter.}

\item{cos_sim_psh}{a vector of size \code{n/2} with the cosine
similarities in high-dimension for the polysphere \eqn{(\mathcal{S}^p)^r}.
The way that the cosine similarities matrix is treated makes the calculus
faster since it is flat in a vector object. Optional parameter, defaults to
\code{NULL}.}
}
\value{
An array of size \code{c(n, n)} with the high-dimension conditional
probabilities of \code{x}.
}
\description{
Calculates the high-dimension conditional probabilities of a
polyspherical Cauchy distribution. Matrix version algorithm.
}
\examples{
x <- sphunif::r_unif_sph(100, 3, 3)
high_dimension(x, rep(0.5, 100))
high_dimension(x, rep(0.5, 100),
               drop(sphunif::Psi_mat(x, scalar_prod = TRUE)))
}
