% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dimension.R
\name{high_dimension_p}
\alias{high_dimension_p}
\title{Scalar version
Calculates the high-dimension probabilities of a polyspherical Cauchy distribution}
\usage{
high_dimension_p(x, rho_list)
}
\arguments{
\item{x}{array 3-dimensional}

\item{rho_list}{rho parameter for each \code{i}-th observation}
}
\value{
a 3-dimensional matrix with the high-dimension probabilities of the array \code{x}
}
\description{
Scalar version
Calculates the high-dimension probabilities of a polyspherical Cauchy distribution
}
\examples{
x <- gen_polysphere(100, 2, 3)
high_dimension_p(x, rep(0.5, 100))
}
