% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_distribution.R
\name{jcondi_psc}
\alias{jcondi_psc}
\title{Calculate the conditional high-dimension probability of the j-th observation given the i-th.}
\usage{
jcondi_psc(x, i, j, rho_list, total_P_i)
}
\arguments{
\item{x}{3d-array that is a poly-sphere, (S^p)^r}

\item{i}{corresponds to the i-th observation index}

\item{j}{corresponds to the j-th observation index}

\item{rho_list}{param values between 0 and 1 (not included)}

\item{total_P_i}{marginal probability of the i-th observation}
}
\value{
conditional poly-spherical Cauchy probability of the j-th given the i-th observation
}
\description{
Calculate the conditional high-dimension probability of the j-th observation given the i-th.
}
\examples{
x <- gen_polysphere(20, 2, 4)
jcondi_psc(x, 1, 2, rep(0.5, 20), P_total_psc(x, rep(0.9999, 20))[1])
jcondi_psc(x, 4, 6, rep(0.9999, 20), P_total_psc(x, rep(0.9999, 20))[4])
}
