% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_distribution.R
\name{jcondi_sc}
\alias{jcondi_sc}
\title{Calculate the probability of choosing a pair of elements
where the i-th and the j-th observations are selected}
\usage{
jcondi_sc(y, i, j, rho, total_Q_i = NULL)
}
\arguments{
\item{y}{matrix that stands for a sphere, S^d}

\item{i}{corresponds to the i-th observation index}

\item{j}{corresponds to the j-th observation index}

\item{rho}{param between 0 and 1 (not included)}

\item{total_Q_i}{is the marginal probability of the i-th observation}
}
\value{
spherical marginal Cauchy density probability given the parameters
}
\description{
Calculate the probability of choosing a pair of elements
where the i-th and the j-th observations are selected
}
\examples{
y <- rotasym::r_unif_sphere(100, 2)
jcondi_sc(y, 1, 3, 0.5)
jcondi_sc(y, 4, 5, 0.9999, Q_i_sc(y, 4, 0.9999))
}
