% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_distribution.R
\name{jcondi_sph}
\alias{jcondi_sph}
\title{Conditional spherical Cauchy probability}
\usage{
jcondi_sph(y, i, j, rho, d_i_sph_cauchy = NULL)
}
\arguments{
\item{y}{a matrix of size \code{c(n, d+1)} with observations on \eqn{\mathcal{S}^d}.}

\item{i}{corresponds to the \eqn{i}-th observation index.}

\item{j}{corresponds to the \eqn{j}-th observation index.}

\item{rho}{concentration parameter must be in [0, 1).}

\item{d_i_sph_cauchy}{the marginal probability of the \eqn{i}-th observation.
Optional, defaults to \code{NULL}.}
}
\value{
Spherical marginal Cauchy density probability given the parameters.
}
\description{
Calculate the probability of choosing a pair of elements
where the \eqn{i}-th and the \eqn{j}-th observations are selected.
}
