% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_dimension.R
\name{low_dimension}
\alias{low_dimension}
\title{Low-dimension probabilities}
\usage{
low_dimension(Y, rho)
}
\arguments{
\item{Y}{matrix of size \code{c(n, d)}, where \code{n} is the number of
observation, with the points onto the sphere \eqn{\mathcal{S}^d}.}

\item{rho}{concentration parameter must be in [0, 1).}
}
\value{
A matrix with the values of x projected onto the sphere of radius 1.
}
\description{
Calculate the low-dimension probabilities of a reduced matrix Y.
}
\examples{
Y <- rotasym::r_unif_sphere(100, 2)
low_dimension(Y, 0)
low_dimension(Y, 0.5)
low_dimension(Y, 0.9999)
}
