% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_distribution.R
\name{P_i_psc}
\alias{P_i_psc}
\title{Calculate the high-dimension poly-spherical Cauchy density for the
i-th observation.}
\usage{
P_i_psc(x, i, rho_list)
}
\arguments{
\item{x}{3d-array that is a poly-sphere, (S^p)^r}

\item{i}{corresponds to the i-th observation index}

\item{rho_list}{param values between 0 and 1 (not included)}
}
\value{
poly-spherical Cauchy density value given the parameters
}
\description{
Calculate the high-dimension poly-spherical Cauchy density for the
i-th observation.
}
\examples{
x <- gen_polysphere(20, 2, 4)
P_i_psc(x, 1, rep(0.5, 20))
P_i_psc(x, 4, rep(0.9999, 20))
}
