% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_distribution.R
\name{P_ij_psc}
\alias{P_ij_psc}
\title{Calculate the high-dimension poly-spherical Cauchy density for the
i-th and j-th observations.}
\usage{
P_ij_psc(x, i, j, rho)
}
\arguments{
\item{x}{3d-array that is a poly-sphere, (S^p)^r}

\item{i}{corresponds to the i-th observation index}

\item{j}{corresponds to the j-th observation index}

\item{rho}{param between 0 and 1 (not included)}
}
\value{
poly-spherical Cauchy density value given the parameters
}
\description{
Calculate the high-dimension poly-spherical Cauchy density for the
i-th and j-th observations.
}
\examples{
x <- gen_polysphere(20, 2, 4)
P_ij_psc(x, 1, 2, 0.5)
P_ij_psc(x, 4, 3, 0.9999)
}
