% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_distribution.R
\name{P_total_psc}
\alias{P_total_psc}
\title{Calculate the marginal high-dimension poly-spherical Cauchy probabilities for the i-th observation.}
\usage{
P_total_psc(x, rho_list)
}
\arguments{
\item{x}{3d-array that is a poly-sphere, (S^p)^r}

\item{rho_list}{param values between 0 and 1 (not included)}
}
\value{
marginal poly-spherical Cauchy probabilities value given the parameters
}
\description{
Calculate the marginal high-dimension poly-spherical Cauchy probabilities for the i-th observation.
}
\examples{
x <- gen_polysphere(20, 2, 4)
P_total_psc(x, rep(0.5, 20))
P_total_psc(x, rep(0.9999, 20))
}
