% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kms.R
\name{plot_kde}
\alias{plot_kde}
\title{Circular kernel mean shift plot}
\usage{
plot_kde(x, h, tol = 0.1, init_clusters = NULL, step = 0.01, cut_tree = TRUE)
}
\arguments{
\item{x}{a matrix of size \code{c(nx, d + 1)} with the initial points.}

\item{h}{defaults to \code{NULL}. If not provided, the kernel mean shift is
calculated with h equal to \code{ks::hpi(x, deriv.order = 1)}.}

\item{tol}{tolerance for equality of modes. Defaults to \code{1e-1}.}

\item{init_clusters}{an array with the label associated to each observation.
Defaults to \code{NULL}.}

\item{step}{separation between evaluation points. Defaults to \code{0.01}.}

\item{cut_tree}{boolean value that informs whether the tree can be cut or not.}
}
\value{
Same object that the function \code{kms_dir} returns.
Defaults to \code{TRUE}.
}
\description{
Produces a circular kernel mean shift plot for the given parameters.
}
\examples{

set.seed(123)
r_mvmf_three <- movMF::rmovMF(n,
                              c(100, 150, 200) * rbind(c(1, 0), c(DirStats::to_cir(pi/2)), c(DirStats::to_cir(pi/4))),
                              c(0.33, 0.33, 0.34))
h_three <- pscsne::bw_kms(r_mvmf_three)
pscsne::plot_kde(r_mvmf_three, h_three)
}
