% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_distribution.R
\name{prob_cond_i_sph}
\alias{prob_cond_i_sph}
\title{Conditional spherical Cauchy probabilities given the
\eqn{i}-th observation}
\usage{
prob_cond_i_sph(y, i, rho)
}
\arguments{
\item{y}{matrix that stands for a sphere, \eqn{\mathcal{S}^d}.}

\item{i}{corresponds to the \eqn{i}-th observation index.}

\item{rho}{concentration parameter must be in [0, 1).}
}
\value{
Spherical marginal Cauchy density probability given the parameters.
}
\description{
Calculate \eqn{n}, where it stands for the sample size,
probabilities where each element is the probability of choosing the
\eqn{i}-th and the \eqn{j}-th observations, the \eqn{i}-th element is a
fixed element in the chosen pairs.
}
\examples{
y <- rotasym::r_unif_sphere(100, 2)
prob_cond_i_sph(y, 1, 0.5)
prob_cond_i_sph(y, 4, 0.9999)
}
