% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_distribution.R
\name{psc_cond_given_i}
\alias{psc_cond_given_i}
\title{Calculate the conditional high-dimension probability for all the j-th observation given the i-th.}
\usage{
psc_cond_given_i(x, i, rho_list, total_P_i = NULL)
}
\arguments{
\item{x}{3d-array that is a poly-sphere, (S^p)^r}

\item{i}{corresponds to the i-th observation index}

\item{rho_list}{param values between 0 and 1 (not included)}

\item{total_P_i}{marginal probability of the i-th observation}
}
\value{
conditional poly-spherical Cauchy probability for all the j-th given the i-th observation
}
\description{
Calculate the conditional high-dimension probability for all the j-th observation given the i-th.
}
\examples{
x <- gen_polysphere(20, 2, 4)
psc_cond_given_i(x, 1, rep(0.5, 20), P_total_psc(x, rep(0.5, 20))[1])
psc_cond_given_i(x, 4, rep(0.9999, 20), P_total_psc(x, rep(0.9999, 20))[4])
}
