% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscsne-package.R
\docType{package}
\name{pscsne-package}
\alias{pscsne-package}
\alias{pscsne}
\title{\code{pscsne}: Polyspherical Cauchy SNE}
\description{
Implementation of polyspherical Cauchy Stochastic Neighbor
Embedding, psc-SNE, a nonlinear dimensionality reduction technique for
polyspherical data (e.g., circular, spherical or toroidal data).
The main function is \code{psc_sne()}.
}
\author{
Eduardo García-Portugués and Luis Ángel Rodríguez García.
}
