% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_distribution.R
\name{Q_i_sc}
\alias{Q_i_sc}
\title{Calculate the marginal low-dimension spherical Cauchy probability}
\usage{
Q_i_sc(y, i, rho)
}
\arguments{
\item{y}{matrix that stands for an sphere, S^d}

\item{i}{corresponds to the i-th observation index}

\item{rho}{param between 0 and 1 (not included)}
}
\value{
spherical marginal Cauchy density probability given the parameters
}
\description{
Calculate the marginal low-dimension spherical Cauchy probability
}
\examples{
y <- rotasym::r_unif_sphere(100, 2)
Q_i_sc(y, 1, 0.5)
Q_i_sc(y, 4, 0.9999)
}
