% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{r_block}
\alias{r_block}
\title{Sample correlation matrices with block structure}
\usage{
r_block(n, g = 5, p = 20, rho = rep(0.9, times = g))
}
\arguments{
\item{n}{sample size.}

\item{g}{number of groups of variables that are uncorrelated between them.}

\item{p}{number of variables on each group.}

\item{rho}{a vector of size \code{g} with the correlations determining the
\code{\link[=stats]{toeplitz}} correlation matrices of each group.}
}
\value{
A matrix of size \code{c(n, g * p)} with the sample.
}
\description{
Sample a zero-mean multivariate normal
\eqn{N_{gp}(\mathbf{0}, \boldsymbol{\Sigma})} with a diagonal block matrix
\eqn{\boldsymbol{\Sigma}} partitioned into \eqn{g} blocks with \eqn{p}
variables. Each block is constructed as a \code{\link[=stats]{toeplitz}}
correlation matrix.
}
\examples{
# Visualize some features
n <- 200
g <- 10
p <- 10
x <- r_block(n = n, g = g, p = p)
pairs(x[, c(1:2, p + 1:2)],
labels = c("Var 1", "Var 2", paste("Var", p + 1), paste("Var", p + 2)))

# Standardize variables -- now the vectors of observations for each variable
# (the columns) live on \sqrt{n - 1} * S^{n - 1}!
x_sca <- scale(x)
colSums(x_sca^2)

# Make the features live on S^{n - 1}
x_sca <- scale(x) / sqrt(n - 1)

# Transpose matrix (features become observations)
feat <- t(x_sca)

# Run psc_sne() with colors being the groups of variables
# dim(feat) <- c(dim(feat), 1)
# cols <- rep(1:g, each = p)
# psc_sne(X = feat, d = 1, colors = cols)
}
