% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{radial_projection}
\alias{radial_projection}
\title{Radial projection onto the sphere}
\usage{
radial_projection(y)
}
\arguments{
\item{y}{matrix with the points in the sphere.}
}
\value{
a matrix with the values of x projected onto the sphere of radius 1.
}
\description{
Projection of the points onto the sphere of radius 1.
}
\examples{
y <- rotasym::r_unif_sphere(100, 2)
radial_projection(y)
}
