% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reconstruct_cos_sim_mat}
\alias{reconstruct_cos_sim_mat}
\title{Reconstruct cosine similarities matrix}
\usage{
reconstruct_cos_sim_mat(cos_sim_vec, n)
}
\arguments{
\item{cos_sim_vec}{vector containing all non-repeated values of the cosine similarity matrix (upper triangular matrix).}

\item{n}{sample size of the original data.}
}
\value{
Cosine similarity matrix formed by the \code{cos_sim_vector}.
}
\description{
Reconstruct the cosine similarities from vector elements to symmetric matrix.
}
\examples{
n <- 6
x <- sphunif::r_unif_sph(n, 3)
cos_sim_vec <- sphunif::Psi_mat(x, scalar_prod = TRUE)
reconstruct_cos_sim_mat(cos_sim_vec[, 1], n)
}
