% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexity.R
\name{rho_optim_ineff}
\alias{rho_optim_ineff}
\title{Serial matrix version
Calculate the rho list values based on a fixed perplexity and a given data in (S^p)^r
Optimize the value using the method L-BFGS-B.
The boundaries are set from 0 to 0.9999. It prints the time consumption.}
\usage{
rho_optim_ineff(x, perplexity)
}
\arguments{
\item{x}{3d-array that is a poly-sphere}

\item{perplexity}{a fixed value to optimize the rho values}
}
\value{
rho list with the values optimized
}
\description{
Serial matrix version
Calculate the rho list values based on a fixed perplexity and a given data in (S^p)^r
Optimize the value using the method L-BFGS-B.
The boundaries are set from 0 to 0.9999. It prints the time consumption.
}
\examples{
x <- gen_polysphere(20, 2, 4)
rho_optim_ineff(x, 22)
rho_optim_ineff(x, 30)
}
