% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexity.R
\name{rho_optim_par}
\alias{rho_optim_par}
\title{Concurrent optimization of the \eqn{\rho} concentration parameters (scalar version)}
\usage{
rho_optim_par(x, perplexity, num_cores = parallel::detectCores() - 1)
}
\arguments{
\item{x}{an array of size \code{c(n, d + 1, r)} with the polyspherical data, where \code{n} is the number of observations, \code{d} is the dimension of each sphere, and \code{r} is the number of spheres.}

\item{perplexity}{a fixed value (between 5 and 100) to optimize the rho parameters.}

\item{num_cores}{number of cores to execute the code concurrently. This value must be below the total number of the CPU has available.}
}
\value{
Rho list (\eqn{\boldsymbol{\rho}}) with the values optimized for the given perplexity.
}
\description{
Calculate the rho list values based on a fixed perplexity and a given data in \eqn{(\mathcal{S}^p)^r}.
Optimize the value using the method L-BFGS-B, setting the boundaries within [0, 1).
The limit is considered as \code{1 - 1e-4}. Each value is calculated concurrently. At the end, the time consumption is shown.
}
\examples{
x <- sphunif::r_unif_sph(20, 3, 4)
rho_optim_par(x, 22, 2)
rho_optim_par(x, 30, 2)
}
