% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexity.R
\name{rho_optimize_1}
\alias{rho_optimize_1}
\title{Concurrent optimization of the \eqn{\rho} concentration parameters
(matrix version)}
\usage{
rho_optimize_1(x, perplexity, num_cores = parallel::detectCores() - 1,
  cos_sim_psh = NULL)
}
\arguments{
\item{x}{an array of size \code{c(n, d + 1, r)} with the polyspherical data,
where \code{n} is the number of observations, \code{d} is the dimension of
each sphere, and \code{r} is the number of spheres.}

\item{perplexity}{a fixed value (between 5 and 100) to optimize the rho
parameters.}

\item{num_cores}{number of cores to execute the code concurrently. This
value must be below the total number of the CPU has available.}

\item{cos_sim_psh}{a vector of size \code{n/2} with the cosine
similarities in high-dimension for the polysphere \eqn{(\mathcal{S}^p)^r}.
The way that the cosine similarities matrix is treated makes the calculus
faster since it is flat in a vector object. Optional parameter, defaults to
\code{NULL}.}
}
\value{
Rho list (\eqn{\boldsymbol{\rho}}) with the values optimized
for the given perplexity.
}
\description{
Calculate the rho list values based on a fixed perplexity and
a given data in \eqn{(\mathcal{S}^p)^r}. Optimize the value using the
method L-BFGS-B. The boundaries are set from 0 to 0.9999. It prints the
time consumption.
}
