% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_distribution.R
\name{sc_cond_given_i}
\alias{sc_cond_given_i}
\title{Calculate n probabilities where each element is the probability of choosing the
i-th and the j-th observations, the i-th element is a fixed element in the chosen pairs.}
\usage{
sc_cond_given_i(y, i, rho)
}
\arguments{
\item{y}{matrix that stands for a sphere, S^d}

\item{i}{corresponds to the i-th observation index}

\item{rho}{param between 0 and 1 (not included)}
}
\value{
spherical marginal Cauchy density probability given the parameters
}
\description{
Calculate n probabilities where each element is the probability of choosing the
i-th and the j-th observations, the i-th element is a fixed element in the chosen pairs.
}
\examples{
y <- rotasym::r_unif_sphere(100, 2)
sc_cond_given_i(y, 1, 0.5)
sc_cond_given_i(y, 4, 0.9999)
}
