% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_distribution.R
\name{simple_dspcauchy_hd}
\alias{simple_dspcauchy_hd}
\title{Calculate the high-dimension spherical Cauchy density function}
\usage{
simple_dspcauchy_hd(x, i, j, rho, k, p)
}
\arguments{
\item{x}{3d-array that is a poly-sphere, (S^p)^r}

\item{i}{corresponds to the i-th observation index}

\item{j}{corresponds to the j-th observation index}

\item{rho}{param between 0 and 1 (not included)}

\item{k}{corresponds to the k-th sphere index}

\item{p}{corresponds to S^p, associated to R^(p+1)}
}
\value{
spherical Cauchy density value given the parameters
}
\description{
Calculate the high-dimension spherical Cauchy density function
}
\examples{
x_1 <- gen_polysphere(20, 1, 4)
simple_dspcauchy_hd(x_1, 4, 3, 0.9999, 3, 1)
x_2 <- gen_polysphere(20, 2, 4)
simple_dspcauchy_hd(x_2, 1, 2, 0.5, 2, 2)
}
