% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_distribution.R
\name{simple_dspcauchy_ld}
\alias{simple_dspcauchy_ld}
\title{Calculate the low-dimension spherical Cauchy density function}
\usage{
simple_dspcauchy_ld(y, i, j, rho, d)
}
\arguments{
\item{y}{matrix that stands for an sphere, S^d}

\item{i}{corresponds to the i-th observation index}

\item{j}{corresponds to the j-th observation index}

\item{rho}{param between 0 and 1 (not included)}

\item{d}{corresponds to S^d, associated to R^(d+1)}
}
\value{
spherical Cauchy density value given the parameters
}
\description{
Calculate the low-dimension spherical Cauchy density function
}
\examples{
y_1 <- rotasym::r_unif_sphere(100, 2)
simple_dspcauchy_ld(y_1, i = 1, j = 2, rho = 0.5, d = 1)
y_2 <- rotasym::r_unif_sphere(100, 3)
simple_dspcauchy_ld(y_2, i = 4, j = 3, rho = 0.9999, d = 2)
}
