% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smallrna}
\alias{smallrna}
\title{Small RNA dataset}
\format{
A data frame with TODO rows and TODO variables:
\describe{
  \item{TODO}{TODO}
  \item{TODO}{TODO}
}
}
\source{
TODO
}
\usage{
smallrna
}
\description{
TODO
}
\details{
TODO
}
\examples{
# Load data
data("smallrna")

# TODO
}
\references{
TODO
}
\keyword{datasets}
