% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dimension.R
\name{symmetric_probs}
\alias{symmetric_probs}
\title{Calculate the symmetric probabilities of a given conditional probabilities matrix}
\usage{
symmetric_probs(P)
}
\arguments{
\item{P}{matrix of probabilities (P_i|j)_{ij}}
}
\value{
The sum of \code{P} and \code{t(P)} divided by twice the sample size
}
\description{
Calculate the symmetric probabilities of a given conditional probabilities matrix
}
\examples{
symmetric_probs(matrix(runif(3 * 3), nrow = 3, ncol = 3))
symmetric_probs(diag(3))
}
