% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexity.R
\name{to_perp}
\alias{to_perp}
\title{Matrix version
Calculate the perplexity of all the observations for a given rho list parameter}
\usage{
to_perp(x, rho_list, cos_sim_ps = NULL)
}
\arguments{
\item{x}{3d-array that is a poly-sphere}

\item{rho_list}{param between 0 and 1 (not included)}

\item{cos_sim_ps}{optional parameter with the cosine similarities of each sphere}
}
\value{
perplexity list for each observation
}
\description{
Matrix version
Calculate the perplexity of all the observations for a given rho list parameter
}
\examples{
x <- gen_polysphere(25, 2, 3)
to_perp(x, rep(0.5, 25))
to_perp(x, rep(0.9999, 25), cosine_polysph(x))
}
