% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexity.R
\name{to_perp_i}
\alias{to_perp_i}
\title{Perplexity for the i-th observation (vector version)}
\usage{
to_perp_i(x, i, rho, cos_sim_psh = NULL)
}
\arguments{
\item{x}{an array of size \code{c(n, d + 1, r)} with the polyspherical data, where \code{n} is the number of observations, \code{d} is the dimension of each sphere, and \code{r} is the number of spheres.}

\item{i}{corresponds to the \eqn{i}-th observation index.}

\item{rho}{concentration parameter must be in [0, 1).}

\item{cos_sim_psh}{cosine similarities array of dimension \code{c(n, n, r)} for the polysphere.}
}
\value{
Perplexity of the \eqn{i}-th observation for all the remainder observations.
}
\description{
Calculate the perplexity of the i-th observation for a given \eqn{\mathbf{\rho}} parameters list. Vector version algorithm.
}
\examples{
x <- sphunif::r_unif_sph(25, 3, 3)
i <- 1
to_perp_i(x, i, 0.25)
}
