% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexity.R
\name{to_perp_scalar}
\alias{to_perp_scalar}
\title{Scalar version
Calculate the perplexity of all the observations for a given rho list parameter}
\usage{
to_perp_scalar(x, rho_list)
}
\arguments{
\item{x}{3d-array that is a poly-sphere}

\item{rho_list}{list param between 0 and 1 (not included) each element}
}
\value{
perplexity list for each observation
}
\description{
Scalar version
Calculate the perplexity of all the observations for a given rho list parameter
}
\examples{
x <- gen_polysphere(25, 2, 3)
to_perp_scalar(x, rep(0.5, 25))
to_perp_scalar(x, rep(0.9999, 25))
}
