% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexity.R
\name{to_perplexity}
\alias{to_perplexity}
\title{Matrix version
Calculate the perplexity of the i-th observation for a given a rho parameter}
\usage{
to_perplexity(x, i, rho, cos_sim_ps = NULL)
}
\arguments{
\item{x}{3d-array that is a poly-sphere}

\item{i}{corresponds to the i-th observation for the perplexity is calculated}

\item{rho}{param between 0 and 1 (not included)}

\item{cos_sim_ps}{cosine similarities of the poly-sphere}
}
\value{
perplexity of the i-th observation for all the j-th's
}
\description{
Matrix version
Calculate the perplexity of the i-th observation for a given a rho parameter
}
\examples{
x <- gen_polysphere(25, 2, 3)
to_perplexity(x, 1, 0.5)
to_perplexity(x, 4, 0.9999)
}
