% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexity.R
\name{to_perplexity_P}
\alias{to_perplexity_P}
\title{Perplexity of the \eqn{i}-th observation (scalar version)}
\usage{
to_perplexity_P(x, i, rho)
}
\arguments{
\item{x}{an array of size \code{c(n, d + 1, r)} with the poly-spherical data, where \code{n} is the number of observations, \code{d} is the dimension of each sphere, and \code{r} is the number of spheres.}

\item{i}{corresponds to the i-th observation for the perplexity is calculated.}

\item{rho}{concentration parameter between 0 and 1 (not included).}
}
\value{
perplexity and probabilities of the \eqn{i}-th observation for all the remainder observations.
}
\description{
Calculate the perplexity of the \eqn{i}-th observation for a given a rho parameter.
}
\examples{
x <- sphunif::r_unif_sph(25, 3, 3)
to_perplexity_P(x, 1, 0.5)
to_perplexity_P(x, 4, 1 - 1e-4)
}
