% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vec2matrix}
\alias{vec2matrix}
\title{Reconstruct the symmetric matrix without diagonal}
\usage{
vec2matrix(vec, n, diag_value)
}
\arguments{
\item{vec}{vector containing all non-repeated values of the cosine
similarity matrix (upper triangular matrix).}

\item{n}{sample size of the original data.}

\item{diag_value}{value for the diagonal of the resultant matrix.}
}
\value{
Matrix formed by the \code{vec}.
}
\description{
Convert the vector into a symmetric matrix with diagonal
equal to a specific value.
}
\examples{
n <- 6
x <- sphunif::r_unif_sph(n, 3)
cos_sim_vec <- drop(sphunif::Psi_mat(x, scalar_prod = TRUE))
vec2matrix(cos_sim_vec, n, diag_value = 1)
}
\keyword{internal}
