% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psc_sne.R
\name{visualize_iter_sol}
\alias{visualize_iter_sol}
\title{Visualize the reduction status}
\usage{
visualize_iter_sol(Y, i, d, colors = NULL)
}
\arguments{
\item{Y}{data point onto the low-dimension \eqn{\mathcal{S}^d}.}

\item{i}{the \eqn{i}-th iteration.}

\item{d}{the dimension to reduce the original data.}

\item{colors}{optional value to represent the group colors in the plot.}
}
\description{
Visualize the iteration solution in a plot for the \eqn{i}-th iteration.
}
