% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexity.R
\name{bin_search}
\alias{bin_search}
\title{Binary search tree algorithm
Calculate the rho value based on a perplexity difference
and current values of rho, rho min and rho max.}
\usage{
bin_search(perp_diff, rho, rho_min, rho_max)
}
\arguments{
\item{perp_diff}{difference between the value obtained and the fixed perplexity}

\item{rho}{the current value optimized}

\item{rho_min}{min value for the current rho}

\item{rho_max}{max value for the current rho}
}
\value{
rho found in this step, currently min and max rho
}
\description{
Binary search tree algorithm
Calculate the rho value based on a perplexity difference
and current values of rho, rho min and rho max.
}
