% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diag_3d}
\alias{diag_3d}
\title{Set each object's diagonal of a 3d-array}
\usage{
diag_3d(x, k, val)
}
\arguments{
\item{x}{an array of size \code{c(n, d + 1, r)} with the polyspherical data, where \code{n} is the number of observations, \code{d} is the dimension of each sphere, and \code{r} is the number of spheres.}

\item{k}{the \code{k}-th sphere from \eqn{(\mathcal{S}^p)^r}.}

\item{val}{value to set to each array's diagonal of the poly-sphere \code{x}.}
}
\value{
The \code{k}-th matrix of \code{x} with the diagonal set to \code{val}.
}
\description{
Set a specific value to the diagonal of a each matrix of a 3 dimensional array.
}
\examples{
x <- sphunif::r_unif_sph(100, 3, 3)
diag_3d(x, 1, 0)
diag_3d(x, 3, 0)
}
