% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fibonacci_lattice}
\alias{fibonacci_lattice}
\title{Generated optimal evenly separated points onto the sphere S^2}
\usage{
fibonacci_lattice(n)
}
\arguments{
\item{n}{number of points to be generated}
}
\value{
evenly optimal separated points onto the low-dimension sphere S^2
}
\description{
Generated optimal evenly separated points onto the sphere S^2
}
\examples{
fibonacci_lattice(100)
fibonacci_lattice(250)
}
