% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_opt_sphere}
\alias{gen_opt_sphere}
\title{Generated optimal evenly separated points onto the sphere S^d}
\usage{
gen_opt_sphere(n, d)
}
\arguments{
\item{n}{number of points to be generated}

\item{d}{reduced dimension}
}
\value{
evenly optimal separated points onto the low-dimension sphere S^d
}
\description{
Generated optimal evenly separated points onto the sphere S^d
}
\examples{
gen_opt_sphere(100, 1)
gen_opt_sphere(250, 2)
}
