% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dimension.R
\name{high_dimension}
\alias{high_dimension}
\title{Matrix version
Calculates the high-dimension probabilities of a polyspherical cauchy distribution}
\usage{
high_dimension(x, rho_list, cos_sim_pol = NULL)
}
\arguments{
\item{x}{array 3-dimensional}

\item{rho_list}{rho parameter for each \code{i}-th observation}

\item{cos_sim_pol}{cosine similarities of the polysphere}
}
\value{
a 3-dimensional array with the high-dimension probabilities of the array \code{x}
}
\description{
Matrix version
Calculates the high-dimension probabilities of a polyspherical cauchy distribution
}
\examples{
x <- gen_polysphere(100, 2, 3)
high_dimension(x, rep(0.5, 100))
high_dimension(x, rep(0.5, 100), cosine_polysph(x))
}
