% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dimension.R
\name{high_dimension_p}
\alias{high_dimension_p}
\title{Polyspherical Cauchy conditional probability matrix (scalar version)}
\usage{
high_dimension_p(x, rho_list)
}
\arguments{
\item{x}{an array of size \code{c(n, d + 1, r)} with the polyspherical data, where \code{n} is the number of observations, \code{d} is the dimension of each sphere, and \code{r} is the number of spheres.}

\item{rho_list}{rho list of size \code{n} for each \code{i}-th observation that stands for the concentration parameter.}
}
\value{
An array of size \code{c(n, n)} with the high-dimension conditional probabilities of \code{x}.
}
\description{
Calculates the high-dimension conditional probabilities of a polyspherical Cauchy distribution. Scalar version algorithm.
}
\examples{
x <- sphunif::r_unif_sph(100, 3, 3)
high_dimension_p(x, rep(0.5, 100))
}
