% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jdf}
\alias{jdf}
\title{Juan de Fuca currents}
\format{
A data frame with TODO rows and TODO variables:
\describe{
  \item{TODO}{TODO}
  \item{TODO}{TODO}
}
}
\source{
TODO
}
\usage{
jdf
}
\description{
TODO
}
\details{
TODO
}
\examples{
# Load data
data("jdf")

# TODO
}
\references{
TODO
}
\keyword{datasets}
