% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psc_sne.R
\name{kl_divergence_grad}
\alias{kl_divergence_grad}
\title{Kullback--Leibler divergence gradient}
\usage{
kl_divergence_grad(Y, i, rho, d, P, cos_sim = NULL, Q = NULL)
}
\arguments{
\item{Y}{matrix of size \code{c(n, d)}, where \code{n} is the number of
observation, with the points onto the sphere \eqn{\mathcal{S}^d}.}

\item{i}{corresponds to the \eqn{i}-th observation index.}

\item{rho}{concentration parameter must be in [0, 1).}

\item{d}{target dimension to reduce the data.}

\item{P}{matrix of size \code{c(n, n)} with the high-dimensional
polyspherical Cauchy probabilities.}

\item{cos_sim}{a vector of size \code{n/2} with the cosine
similarities in high-dimension for the polysphere \eqn{(\mathcal{S}^p)^r}.
The way that the cosine similarities matrix is treated makes the calculus
faster since it is flat in a vector object. Optional parameter, defaults to
\code{NULL}.}

\item{Q}{matrix of size \code{c(n, n)} with the low-dimension spherical
Cauchy probabilities. Optional parameter, defaults to \code{NULL}.}
}
\value{
Resulting reduced data for the \eqn{i}-th observation onto the
sphere \eqn{\mathcal{S}^d}.
}
\description{
Calculates analytically the gradient of the Kullback--Leibler
divergence between low- and high-dimensional pairwise probabilities.
}
\examples{
Y <- sphunif::r_unif_sph(40, 3, 1)[ , , 1]
X <- sphunif::r_unif_sph(40, 3, 3)
P <- high_dimension(X, rep(0.5, 40), num_cores = 2)
kl_divergence_grad(Y, 3, 0.5, 2, P)
cos_sim <- vec2matrix(
    vec = drop(sphunif::Psi_mat(array(X, dim = c(nrow(X), ncol(X), 1)),
                                   scalar_prod = TRUE)),
    n = nrow(X),
    diag_value = 1)
Q <- low_dimension(Y, 0.5)
kl_divergence_grad(Y, 3, 0.5, 2, P, cos_sim, Q)
}
