% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kms.R
\name{kms_linear}
\alias{kms_linear}
\title{Kernel mean shift clustering for linear data}
\usage{
kms_linear(x, original_clusters, h = ks::hpi(x, deriv.order = 1))
}
\arguments{
\item{x}{a matrix of size \code{c(nx, d + 1)} with the initial points.}

\item{original_clusters.}{Vectors with the label associated to each observation.}

\item{h.}{Defaults to \code{ks::hpi(x, deriv.order = 1)}.}
}
\value{
A list with the following entries:
\itemize{
  \item \code{cluster}: vector giving the cluster labels.
  \item \code{unique_modes}: estimated modes for each cluster (sorted).
  \item \code{h}: used bandwidth.
  \item \code{position_antimodes}: indexes with the location of the antimodes.
  \item \code{labels_rle_values}: values of the label's rle
  \item \code{x_kms}: eval points
}
}
\description{
Computes the kernel mean shift clustering for scalar values.
}
