% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kms.R
\name{plot_kde}
\alias{plot_kde}
\title{Plot the density chart of the linear data variable}
\usage{
plot_kde(x, original_clusters, h = ks::hpi(x, deriv.order = 1))
}
\arguments{
\item{x}{a matrix of size \code{c(nx, d + 1)} with the initial points.}

\item{original_clusters.}{Vectors with the label associated to each observation.}

\item{h.}{Defaults to \code{ks::hpi(x, deriv.order = 1)}.}
}
\description{
Produces a density plot for the given parameters.
}
