% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psc_sne.R
\name{psc_sne}
\alias{psc_sne}
\title{Polyspherical Cauchy SNE}
\usage{
psc_sne(X, d, rho_psc_list = NULL, rho = 0.5, perplexity = 30,
  num_iteration = 200, initial_momentum = 0.5, final_momentum = 0.8,
  eta = 200, early_exaggeration = 4, colors = NULL,
  visualize_prog = FALSE, tol = 1e-09, check = TRUE,
  parallel_cores = parallel::detectCores() - 1)
}
\arguments{
\item{X}{an array of size \code{c(n, d + 1, r)} with the polyspherical data,
where \code{n} is the number of observations, \code{d} is the dimension of
each sphere, and \code{r} is the number of spheres.}

\item{d}{the target dimension to use for reduce the dimension of the data
\code{X}.}

\item{rho_psc_list}{rho parameters of the high-dimensional polyspherical
Cauchy probabilities (optional, default \code{NULL}).}

\item{rho}{rho parameter of the low-dimensional spherical Cauchy
probabilities (optional, default value \code{0.5}).}

\item{perplexity}{parameter which says what is more important: local or
global aspects (optional, default \code{30}).}

\item{num_iteration}{maximum number of iterations (optional, default value
\code{200}).}

\item{initial_momentum}{first value of the momentum of the first \code{250}
iterations.}

\item{final_momentum}{momentum to take into account after the \code{250}
iteration.}

\item{eta}{is the learning rate of the optimization algorithm (optional,
default value \code{200}).}

\item{early_exaggeration}{the first \code{100} iterations results are
exaggerated by this factor (optional, default value \code{4.0}).}

\item{colors}{list with as many elements as observations are, only valid
when visualization is true (optional, default value \code{NULL}).}

\item{visualize_prog}{defines whether the progression plots are shown or
not (optional, default value \code{FALSE}).}

\item{tol}{is the tolerance, when is below this value it is considered that
a good solution has been obtained (optional, default value \code{1e-9}).}

\item{check}{whether to check or not the tolerance.}

\item{parallel_cores}{number of cores to use concurrently for the
calculation of the gradient.}
}
\value{
Resulting data reduced to \eqn{\mathcal{S}^d} after applying the
algorithm for the total number of iterations selected.
}
\description{
Calculates the polyspherical-Cauchy SNE given a data onto the
polysphere and the reduced dimension.
}
\examples{
X <- sphunif::r_unif_sph(40, 3, 3)
psc_sne(X, d = 1, parallel_cores = 2, num_iteration = 1e4)
psc_sne(X, d = 2, parallel_cores = 2)
}
