% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscsne-package.R
\docType{package}
\name{pscsne-package}
\alias{pscsne-package}
\alias{pscsne}
\title{\code{pscsne}: Polyspherical Cauchy SNE}
\description{
Implementation of the Polyspherical Cauchy SNE
}
\author{
Eduardo García-Portugués and Luis Ángel Rodríguez García.
}
