% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexity.R
\name{rho_optim_i_bst}
\alias{rho_optim_i_bst}
\title{Binary search rho optimization for the \eqn{i}-th observation}
\usage{
rho_optim_i_bst(x, i, perp_fixed, tolerance = 0.001, rho = 0.5, max_tries = 20)
}
\arguments{
\item{x}{an array of size \code{c(n, d + 1, r)} with the polyspherical data, where \code{n} is the number of observations, \code{d} is the dimension of each sphere, and \code{r} is the number of spheres.}

\item{i}{the \eqn{i}-th observation.}

\item{perp_fixed}{a fixed value used to optimized the rho values.}

\item{tolerance}{whether the difference between previous and current results is below this value (optional, default value \code{1e-3}).}

\item{rho}{parameter which determines the concentration of the spherical Cauchy distribution (optional, default value \code{0.5}).}

\item{max_tries}{number of maximum tries for each value of the rho list.}
}
\value{
Rho concentration parameter and conditional probabilities calculated for the \eqn{i}-th observation.
}
\description{
Calculate the rho based on a fixed perplexity and a given data in \eqn{(\mathcal{S}^p)^r}.
The boundaries are set from 0 to 0.9999.
}
