% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexity.R
\name{rho_optimize_1}
\alias{rho_optimize_1}
\title{Parallel matrix version
Calculate the rho list values based on a fixed perplexity and a given data in (S^p)^r
Optimize the value using the method L-BFGS-B.
The boundaries are set from 0 to 0.9999. It prints the time consumption.}
\usage{
rho_optimize_1(x, perplexity, num_cores = parallel::detectCores() - 1)
}
\arguments{
\item{x}{3d-array that is a poly-sphere}

\item{perplexity}{a fixed value to optimize the rho values}

\item{num_cores}{number of cores to execute the code concurrently}
}
\value{
rho list with the values optimized
}
\description{
Parallel matrix version
Calculate the rho list values based on a fixed perplexity and a given data in (S^p)^r
Optimize the value using the method L-BFGS-B.
The boundaries are set from 0 to 0.9999. It prints the time consumption.
}
\examples{
x <- gen_polysphere(20, 2, 4)
rho_optimize_1(x, 22, 2)
rho_optimize_1(x, 30, 2)
}
