% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexity.R
\name{rho_optimize_1}
\alias{rho_optimize_1}
\title{Concurrent optimization of the \eqn{\rho} concentration parameters (matrix version)}
\usage{
rho_optimize_1(x, perplexity, num_cores = parallel::detectCores() - 1)
}
\arguments{
\item{x}{an array of size \code{c(n, d + 1, r)} with the poly-spherical data, where \code{n} is the number of observations, \code{d} is the dimension of each sphere, and \code{r} is the number of spheres.}

\item{perplexity}{a fixed value (between 5 and 100) to optimize the rho values.}

\item{num_cores}{number of cores to execute the code concurrently.}
}
\value{
rho list with the values optimized.
}
\description{
Calculate the rho list values based on a fixed perplexity and a given data in \eqn{(\mathcal{S}^p)^r}.
Optimize the value using the method L-BFGS-B.
The boundaries are set from 0 to 0.9999. It prints the time consumption.
}
