% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexity.R
\name{rho_optimParallel}
\alias{rho_optimParallel}
\title{Parallel scalar version
Calculate the rho list values based on a fixed perplexity and a given data in (S^p)^r
Optimize the value using the concurrently method L-BFGS-B (optimParallel).
The boundaries are set from 0 to 0.9999. It prints the time consumption.}
\usage{
rho_optimParallel(x, perplexity, num_cores = parallel::detectCores() - 1)
}
\arguments{
\item{x}{3d-array that is a poly-sphere}

\item{perplexity}{a fixed value to optimize the rho values}

\item{num_cores}{number of cores to execute the code concurrently}
}
\value{
rho list with the values optimized
}
\description{
Parallel scalar version
Calculate the rho list values based on a fixed perplexity and a given data in (S^p)^r
Optimize the value using the concurrently method L-BFGS-B (optimParallel).
The boundaries are set from 0 to 0.9999. It prints the time consumption.
}
