% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexity.R
\name{to_perp}
\alias{to_perp}
\title{Perplexity matrix (matrix version)}
\usage{
to_perp(x, rho_list, cos_sim_psh = NULL)
}
\arguments{
\item{x}{an array of size \code{c(n, d + 1, r)} with the polyspherical data,
where \code{n} is the number of observations, \code{d} is the dimension of
each sphere, and \code{r} is the number of spheres.}

\item{rho_list}{rho list of size \code{n} for each \eqn{i}-th observation
that stands for the concentration parameter.}

\item{cos_sim_psh}{a vector of size \code{n/2} with the cosine
similarities in high-dimension for the polysphere \eqn{(\mathcal{S}^p)^r}.
The way that the cosine similarities matrix is treated makes the calculus
faster since it is flat in a vector object. Optional parameter, defaults to
\code{NULL}.}
}
\value{
Perplexity of each \eqn{i}-th observation for all the remainder
observations.
}
\description{
Calculate the perplexity of each observations for a given
\eqn{\mathbf{\rho}} parameters list. Matrix version algorithm.
}
\examples{
x <- sphunif::r_unif_sph(25, 3, 3)
to_perp(x, rep(0.5, 25))
to_perp(x, rep(1 - 1e-4, 25), sphunif::Psi_mat(x, scalar_prod = TRUE))
}
