% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexity.R
\name{to_perp_i}
\alias{to_perp_i}
\title{Perplexity for the i-th observation}
\usage{
to_perp_i(x, i, rho, cos_sim_psh = NULL)
}
\arguments{
\item{x}{an array of size \code{c(n, d + 1, r)} with the polyspherical data,
where \code{n} is the number of observations, \code{d} is the dimension of
each sphere, and \code{r} is the number of spheres.}

\item{i}{corresponds to the \eqn{i}-th observation index.}

\item{rho}{concentration parameter must be in [0, 1).}

\item{cos_sim_psh}{a vector of size \code{n/2} with the cosine
similarities in high-dimension for the polysphere \eqn{(\mathcal{S}^p)^r}.
The way that the cosine similarities matrix is treated makes the calculus
faster since it is flat in a vector object. Optional parameter, defaults to
\code{NULL}.}
}
\value{
Perplexity of the \eqn{i}-th observation for all
the remainder observations.
}
\description{
Calculate the perplexity of the i-th observation for a given
\eqn{\boldsymbol{\rho}} parameters list.
}
\examples{
x <- sphunif::r_unif_sph(25, 3, 3)
i <- 1
to_perp_i(x, i, 0.25)
}
