% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psc_sne.R
\name{visualize_iter_sol}
\alias{visualize_iter_sol}
\title{Visualize the reduction status}
\usage{
visualize_iter_sol(Y, i, d, colors = NULL)
}
\arguments{
\item{Y}{matrix of size \code{c(n, d)}, where \code{n} is the number of observation, with the points onto the sphere \eqn{\mathcal{S}^d}.}

\item{i}{the \eqn{i}-th iteration.}

\item{d}{the dimension to reduce the original data.}

\item{colors}{defines the group colors in the plot. Optional parameter,
default value set to \code{NULL}.}
}
\description{
Visualize the iteration solution in a plot for the
\eqn{i}-th iteration.
}
